more magic!
